+++ docs/reference/changelog.md
@@ -22,6 +22,7 @@ This is the pre-release series. Things haven't settled yet, so each minor versio
 - Handle activities without any location points. [GH-10](https://github.com/martin-ueding/geo-activity-playground/issues/10)
 - Resolve Strava Gear name. [GH-18](https://github.com/martin-ueding/geo-activity-playground/issues/18)
 - Add page for equipment. [GH-3](https://github.com/martin-ueding/geo-activity-playground/issues/3)
+- Add a pop-up with some metadata about the first visit to the explorer tiles. [GH-14](https://github.com/martin-ueding/geo-activity-playground/issues/14)
 
 ### Version 0.5